requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.5
djangorestframework==3.15.2
filelock==3.16.1
numpy==2.1.3
platformdirs==4.3.6
sqlparse==0.5.3
tzdata==2024.2
virtualenv==20.28.1
